+++ gspread/models.py
@@ -13,7 +13,7 @@ try:
 except:
     from urllib import quote
 
-from .base import BaseCell, BaseSpreadsheet
+from .base import BaseCell
 
 from .exceptions import WorksheetNotFound, CellNotFound
 
@@ -41,7 +41,7 @@ except NameError:
     basestring = unicode = str
 
 
-class Spreadsheet(BaseSpreadsheet):
+class Spreadsheet(object):
     """The class that represents a spreadsheet."""
     def __init__(self, client, properties):
         self.client = client
@@ -226,6 +226,65 @@ class Spreadsheet(BaseSpreadsheet):
         }
 
         return self.batch_update(body)
+
+    def share(self, value, perm_type, role, notify=True, email_message=None):
+        """Share the spreadsheet with other accounts.
+        :param value: user or group e-mail address, domain name
+                      or None for 'default' type.
+        :param perm_type: the account type.
+               Allowed values are: ``user``, ``group``, ``domain``,
+               ``anyone``.
+        :param role: the primary role for this user.
+               Allowed values are: ``owner``, ``writer``, ``reader``.
+        :param notify: Whether to send an email to the target user/domain.
+        :param email_message: The email to be sent if notify=True
+
+        Example::
+
+            # Give Otto a write permission on this spreadsheet
+            sh.share('otto@example.com', perm_type='user', role='writer')
+
+            # Transfer ownership to Otto
+            sh.share('otto@example.com', perm_type='user', role='owner')
+
+        """
+        self.client.insert_permission(
+            self.id,
+            value=value,
+            perm_type=perm_type,
+            role=role,
+            notify=notify,
+            email_message=email_message
+        )
+
+    def list_permissions(self):
+        """Lists the spreadsheet's permissions.
+        """
+        return self.client.list_permissions(self.id)
+
+    def remove_permissions(self, value, role='any'):
+        """
+        Example::
+
+            # Remove Otto's write permission for this spreadsheet
+            sh.remove_permissions('otto@example.com', role='writer')
+
+            # Remove all Otto's permissions for this spreadsheet
+            sh.remove_permissions('otto@example.com')
+        """
+        permission_list = self.client.list_permissions(self.id)
+
+        key = 'emailAddress' if '@' in value else 'domain'
+
+        filtered_id_list = [
+            p['id'] for p in permission_list
+            if p[key] == value and (p['role'] == role or role == 'any')
+        ]
+
+        for permission_id in filtered_id_list:
+            self.client.remove_permission(self.id, permission_id)
+
+        return filtered_id_list
 
 
 class Worksheet(object):
